metaclassifier/requirements.txt
# MetaClassifier Requirements

# Core dependencies
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0
biopython>=1.81

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Utilities
scikit-learn>=1.3.0

# Optional: For Evo2 integration
# evo2  # Install from: https://github.com/ArcInstitute/evo2

# Optional: For Excel output
# openpyxl>=3.1.0

# Development (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0

